Rewrite EventBus
Instead of attempting to write all events persistently into a pending
tree, instead maintain an in-memory priority queue of events.

This makes simulation startup significantly faster because the entire
history of daily bars doesn't need to be dropped into the
$sim/events/pending tree--these can be seeded directly from
/history/daily.

This change also includes minor changes to the format of the historical
data so that they can be directly replayed as events. Now that these can
be replayed directly, there's no need for the MarketDataHandler so this
has been removed too.